# public/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.1",
    "title": "Hello World App"
  },
  "host": "localhost:10010",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/movie/{id}": {
      "x-swagger-router-controller": "movie",
      "get": {
        "operationId": "getOne",
        "description": "get details of a single movie",
        "parameters": [
          {
            "name": "id",
            "type": "number",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetMovieResponse"
            }
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "put": {
        "operationId": "update",
        "description": "update movie",
        "parameters": [
          {
            "name": "id",
            "type": "number",
            "in": "path",
            "required": true
          },
          {
            "name": "title",
            "in": "body",
            "required": true,
            "description": "properties of movie",
            "schema": {
              "$ref": "#/definitions/Movie"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GeneralResponse"
            }
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "operationId": "delMovie",
        "description": "delete a movie",
        "parameters": [
          {
            "name": "id",
            "type": "number",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GeneralResponse"
            }
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "GetMovieResponse": {
      "required": [
        "title",
        "year",
        "id"
      ],
      "properties": {
        "title": {
          "type": "string",
          "description": "title of movie"
        },
        "year": {
          "type": "number",
          "description": "year of publish"
        },
        "id": {
          "type": "number"
        }
      }
    },
    "Movie": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "description": "task object name"
        },
        "year": {
          "type": "number",
          "description": "year of release"
        }
      },
      "required": [
        "title",
        "year"
      ]
    },
    "GeneralResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "number",
          "description": "return 1 if successful"
        },
        "description": {
          "type": "string",
          "description": "a short comment"
        }
      },
      "required": [
        "success",
        "description"
      ]
    },
...
    "ErrorResponse": {
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "type": "string"
        }
      }
    }
  }
}